requirements.txt
mistralai>=0.0.7
openai>=1.6.0
python-dotenv>=1.0.0
tqdm>=4.66.1
requests>=2.31.0 